Move attachment serialization helpers to File base class
