+++ src/Files/Base64Document.php
@@ -7,11 +7,10 @@ use Illuminate\Http\UploadedFile;
 use JsonSerializable;
 use Laravel\Ai\Contracts\Files\StorableFile;
 use Laravel\Ai\Files\Concerns\CanBeUploadedToProvider;
-use Laravel\Ai\Files\Concerns\HasStoredBase64;
 
 class Base64Document extends Document implements Arrayable, JsonSerializable, StorableFile
 {
-    use CanBeUploadedToProvider, HasStoredBase64;
+    use CanBeUploadedToProvider;
 
     public function __construct(public string $base64, ?string $mimeType = null)
     {
@@ -35,6 +34,11 @@ class Base64Document extends Document implements Arrayable, JsonSerializable, St
     public function content(): string
     {
         return base64_decode($this->base64);
+    }
+
+    public function base64(): string
+    {
+        return $this->base64;
     }
 
     /**
@@ -70,5 +74,4 @@ class Base64Document extends Document implements Arrayable, JsonSerializable, St
     {
         return $this->content();
     }
-
 }
