+++ src/Files/Base64Image.php
@@ -6,11 +6,10 @@ use Illuminate\Contracts\Support\Arrayable;
 use JsonSerializable;
 use Laravel\Ai\Contracts\Files\StorableFile;
 use Laravel\Ai\Files\Concerns\CanBeUploadedToProvider;
-use Laravel\Ai\Files\Concerns\HasStoredBase64;
 
 class Base64Image extends Image implements Arrayable, JsonSerializable, StorableFile
 {
-    use CanBeUploadedToProvider, HasStoredBase64;
+    use CanBeUploadedToProvider;
 
     public function __construct(public string $base64, ?string $mimeType = null)
     {
@@ -23,6 +22,11 @@ class Base64Image extends Image implements Arrayable, JsonSerializable, Storable
     public function content(): string
     {
         return base64_decode($this->base64);
+    }
+
+    public function base64(): string
+    {
+        return $this->base64;
     }
 
     /**
@@ -58,5 +62,4 @@ class Base64Image extends Image implements Arrayable, JsonSerializable, Storable
     {
         return $this->content();
     }
-
 }
