--- src/Files/Concerns/HasStoredBase64.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace Laravel\Ai\Files\Concerns;
-
-trait HasStoredBase64
-{
-    public function base64(): string
-    {
-        return $this->base64;
-    }
-}
