+++ src/Files/LocalAudio.php
@@ -84,5 +84,4 @@ class LocalAudio extends Audio implements Arrayable, JsonSerializable, StorableF
     {
         return $this->content();
     }
-
 }
