+++ src/Files/LocalImage.php
@@ -73,5 +73,4 @@ class LocalImage extends Image implements Arrayable, JsonSerializable, StorableF
     {
         return $this->content();
     }
-
 }
