+++ src/Files/RemoteDocument.php
@@ -37,5 +37,4 @@ class RemoteDocument extends Document implements Arrayable, JsonSerializable, St
     {
         return $this->toArray();
     }
-
 }
