+++ src/Files/StoredAudio.php
@@ -76,5 +76,4 @@ class StoredAudio extends Audio implements Arrayable, JsonSerializable, Storable
     {
         return $this->content();
     }
-
 }
