+++ src/Files/StoredDocument.php
@@ -65,5 +65,4 @@ class StoredDocument extends Document implements Arrayable, JsonSerializable, St
     {
         return $this->content();
     }
-
 }
