+++ src/Files/StoredImage.php
@@ -65,5 +65,4 @@ class StoredImage extends Image implements Arrayable, JsonSerializable, Storable
     {
         return $this->content();
     }
-
 }
