+++ src/Gateway/Mistral/Concerns/MapsAttachments.php
@@ -7,15 +7,14 @@ use Illuminate\Support\Collection;
 use InvalidArgumentException;
 use Laravel\Ai\Files\Base64Image;
 use Laravel\Ai\Files\File;
+use Laravel\Ai\Files\Image;
 use Laravel\Ai\Files\LocalImage;
 use Laravel\Ai\Files\RemoteDocument;
 use Laravel\Ai\Files\RemoteImage;
 use Laravel\Ai\Files\StoredImage;
-use Laravel\Ai\Gateway\Concerns\MapsUploadedFiles;
 
 trait MapsAttachments
 {
-    use MapsUploadedFiles;
     /**
      * Map the given Laravel attachments to Chat Completions content parts.
      */
@@ -47,7 +46,7 @@ trait MapsAttachments
                 ],
                 $attachment instanceof UploadedFile && $this->isImage($attachment) => [
                     'type' => 'image_url',
-                    'image_url' => ['url' => $this->uploadedFileAsDataUri($attachment)],
+                    'image_url' => ['url' => Image::fromUpload($attachment)->asDataUri()],
                 ],
                 $attachment instanceof RemoteDocument => [
                     'type' => 'document_url',
@@ -73,5 +72,4 @@ trait MapsAttachments
             'image/webp',
         ]);
     }
-
 }
