+++ src/Gateway/OpenAi/Concerns/MapsAttachments.php
@@ -7,7 +7,9 @@ use Illuminate\Support\Collection;
 use InvalidArgumentException;
 use Laravel\Ai\Files\Base64Document;
 use Laravel\Ai\Files\Base64Image;
+use Laravel\Ai\Files\Document;
 use Laravel\Ai\Files\File;
+use Laravel\Ai\Files\Image;
 use Laravel\Ai\Files\LocalDocument;
 use Laravel\Ai\Files\LocalImage;
 use Laravel\Ai\Files\ProviderDocument;
@@ -16,11 +18,9 @@ use Laravel\Ai\Files\RemoteDocument;
 use Laravel\Ai\Files\RemoteImage;
 use Laravel\Ai\Files\StoredDocument;
 use Laravel\Ai\Files\StoredImage;
-use Laravel\Ai\Gateway\Concerns\MapsUploadedFiles;
 
 trait MapsAttachments
 {
-    use MapsUploadedFiles;
     /**
      * Map the given Laravel attachments to OpenAI content parts.
      */
@@ -80,11 +80,11 @@ trait MapsAttachments
                 ],
                 $attachment instanceof UploadedFile && $this->isImage($attachment) => [
                     'type' => 'input_image',
-                    'image_url' => $this->uploadedFileAsDataUri($attachment),
+                    'image_url' => Image::fromUpload($attachment)->asDataUri(),
                 ],
                 $attachment instanceof UploadedFile => [
                     'type' => 'input_file',
-                    'file_data' => $this->uploadedFileAsDataUri($attachment),
+                    'file_data' => Document::fromUpload($attachment)->asDataUri(),
                     'filename' => $attachment->getClientOriginalName(),
                 ],
                 default => throw new InvalidArgumentException('Unsupported attachment type ['.get_class($attachment).']'),
@@ -117,5 +117,4 @@ trait MapsAttachments
             default => '',
         };
     }
-
 }
